requirements.txt
Django==1.8.17
django-role-permissions==2.2.1
datadog==0.37.1
ddtrace==0.39.0
six==1.12.0
future==0.15.2
futures==3.2.0
django-orderable==6.0.1
django-adminplus==0.2.1
